+++ test/mix/tasks/gen_site_test.exs
@@ -2,8 +2,14 @@ defmodule Mix.Tasks.Beacon.GenSiteTest do
   use Beacon.CodeGenCase
   import Igniter.Test
 
+  @secret_key_base "A0DSgxjGCYZ6fCIrBlg6L+qC/cdoFq5Rmomm53yacVmN95Wcpl57Gv0sTJjKjtIo"
+  @signing_salt "O68x1k5B"
+  @port 4041
+  @secure_port 8445
+
   @opts_my_site ~w(--site my_site --path /)
   @opts_other_site ~w(--site other --path /other)
+  @opts_host ~w(--site my_site --path / --host example.com --port #{@port} --secure-port #{@secure_port} --secret-key-base #{@secret_key_base} --signing-salt #{@signing_salt})
 
   describe "options validation" do
     test "validates site" do
@@ -155,6 +161,148 @@ defmodule Mix.Tasks.Beacon.GenSiteTest do
       |> assert_has_patch("lib/test/application.ex", """
       14    - |      {Beacon, [sites: [Application.fetch_env!(:beacon, :my_site)]]},
          14 + |      {Beacon, [sites: [Application.fetch_env!(:beacon, :my_site), Application.fetch_env!(:beacon, :other)]]},
+      """)
+    end
+  end
+
+  describe "--host option" do
+    setup do
+      [project: phoenix_project()]
+    end
+
+    test "creates endpoint", %{project: project} do
+      project
+      |> Igniter.compose_task("beacon.gen.site", @opts_host)
+      |> assert_creates("lib/test_web/example_endpoint.ex", """
+      defmodule TestWeb.ExampleEndpoint do
+        use Phoenix.Endpoint, otp_app: :test
+
+        @session_options Application.compile_env!(:test, :session_options)
+
+        # socket /live must be in the proxy endpoint
+
+        # Serve at "/" the static files from "priv/static" directory.
+        #
+        # You should set gzip to true if you are running phx.digest
+        # when deploying your static files in production.
+        plug Plug.Static,
+          at: "/",
+          from: :test,
+          gzip: false,
+          only: TestWeb.static_paths()
+
+        # Code reloading can be explicitly enabled under the
+        # :code_reloader configuration of your endpoint.
+        if code_reloading? do
+          socket "/phoenix/live_reload/socket", Phoenix.LiveReloader.Socket
+          plug Phoenix.LiveReloader
+          plug Phoenix.CodeReloader
+          plug Phoenix.Ecto.CheckRepoStatus, otp_app: :test
+        end
+
+        plug Plug.RequestId
+        plug Plug.Telemetry, event_prefix: [:phoenix, :endpoint]
+
+        plug Plug.Parsers,
+          parsers: [:urlencoded, :multipart, :json],
+          pass: ["*/*"],
+          json_decoder: Phoenix.json_library()
+
+        plug Plug.MethodOverride
+        plug Plug.Head
+        plug Plug.Session, @session_options
+        plug TestWeb.Router
+      end
+      """)
+    end
+
+    test "updates config.exs", %{project: project} do
+      project
+      |> Igniter.compose_task("beacon.gen.site", @opts_host)
+      # add config for new endpoint
+      |> assert_has_patch("config/config.exs", """
+         10 + |config :test, TestWeb.ExampleEndpoint,
+         11 + |  url: [host: "localhost"],
+         12 + |  adapter: Bandit.PhoenixAdapter,
+         13 + |  render_errors: [
+         14 + |    formats: [html: TestWeb.ErrorHTML, json: TestWeb.ErrorJSON],
+         15 + |    layout: false
+         16 + |  ],
+         17 + |  pubsub_server: Test.PubSub,
+         18 + |  live_view: [signing_salt: "#{@signing_salt}"]
+      """)
+      # update signing salt for host app session_options
+      |> assert_has_patch("config/config.exs", """
+         28 + |    signing_salt: "#{@signing_salt}",
+      """)
+      # update signing salt for existing endpoint
+      |> assert_has_patch("config/config.exs", """
+         41 + |  live_view: [signing_salt: "#{@signing_salt}"]
+      """)
+    end
+
+    test "updates dev.exs", %{project: project} do
+      project
+      |> Igniter.compose_task("beacon.gen.site", @opts_host)
+      # add config for new endpoint
+      |> assert_has_patch("config/dev.exs", """
+         3 + |config :test, TestWeb.ExampleEndpoint,
+         4 + |  http: [ip: {127, 0, 0, 1}, port: #{@port}],
+         5 + |  check_origin: false,
+         6 + |  code_reloader: true,
+         7 + |  debug_errors: true,
+         8 + |  secret_key_base: "#{@secret_key_base}",
+         9 + |  watchers: [
+        10 + |    esbuild: {Esbuild, :install_and_run, [:default, ~w(--sourcemap=inline --watch)]},
+        11 + |    tailwind: {Tailwind, :install_and_run, [:default, ~w(--watch)]}
+        12 + |  ]
+        13 + |
+      """)
+      # update secret key base for existing endpoint
+      |> assert_has_patch("config/dev.exs", """
+          33 + |  secret_key_base: "#{@secret_key_base}",
+      """)
+    end
+
+    test "updates runtime.exs", %{project: project} do
+      project
+      |> Igniter.compose_task("beacon.gen.site", @opts_host)
+      # add beacon site config
+      |> assert_has_patch("config/runtime.exs", """
+         2 + |config :beacon, my_site: [site: :my_site, repo: Test.Repo, endpoint: TestWeb.ExampleEndpoint, router: TestWeb.Router]
+      """)
+      # adds host to proxy endpoint allowed origins
+      |> assert_has_patch("config/runtime.exs", """
+        48 + |    check_origin: [host, "example.com"],
+      """)
+      # add config for new endpoint
+      |> assert_has_patch("config/runtime.exs", """
+        54 + |config :test, TestWeb.ExampleEndpoint,
+        55 + |  url: [host: "example.com", port: #{@secure_port}, scheme: "https"],
+        56 + |  http: [ip: {0, 0, 0, 0, 0, 0, 0, 0}, port: #{@port}],
+        57 + |  secret_key_base: secret_key_base,
+        58 + |  server: !!System.get_env("PHX_SERVER")
+        59 + |
+      """)
+    end
+
+    test "updates application.ex", %{project: project} do
+      project
+      |> Igniter.compose_task("beacon.gen.site", @opts_host)
+      |> assert_has_patch("lib/test/application.ex", """
+        22 + | TestWeb.ExampleEndpoint,
+      """)
+    end
+
+    test "updates router", %{project: project} do
+      project
+      |> Igniter.compose_task("beacon.gen.site", @opts_host)
+      |> assert_has_patch("lib/test_web/router.ex", """
+        23 + |  scope "/", host: ["localhost", "example.com"] do
+        24 + |    pipe_through [:browser, :beacon]
+        25 + |    beacon_site "/", site: :my_site
+        26 + |  end
+        27 + |
       """)
     end
   end